+++ CHANGELOG.md
@@ -4,7 +4,9 @@ Yii Framework 2 mongodb extension Change Log
 2.0.5 under development
 -----------------------
 
+- Enh #23: Added support for complex sort specification at `yii\mongodb\Query` (raoptimus)
 - Enh #2022: Added support for querying on values using basic generic operators to `yii\mongodb\Query`
+
 
 2.0.4 May 10, 2015
 ------------------